+++ src/main.cpp
@@ -20,6 +20,7 @@ class $modify(ModProfilesMenuLayer, MenuLayer) {
 		);
 
 		auto menu = this->getChildByID("bottom-menu");
+		myButton->setScale(0.5f);
 		menu->addChild(myButton);
 
 		myButton->setID("my-button"_spr);