+++ changelog.md
@@ -1,8 +1,11 @@
-# v1.0.2
+# v1.1.0
 * Optimized Mod Downloading Further (No longer lags while downloading)
 * Added Discord Button
+* Added Restart Button
 * Fixed crash while importing mods
 * Fixed errors with importing and exporting
+* Fixed mod file names not being correct
+* Fixed crash while downloading mods
 
 # v1.0.1
 * Optimized Mod Downloading ([ninXout](https://github.com/ninXout))